Fixed DolibarrIntervention date properties types

+++ Sources/SwiftDolibarr/Models/Core/Intervention/DolibarrIntervention.swift
@@ -72,12 +72,12 @@ public final class DolibarrIntervention: CommonBusinessObject {
 	/// Intervention start date (Unix timestamp)
 	///
 	/// - Mapped Dolibarr property: **dateo**
-	public var dateStart: Double?
+	public var dateStart: Int?
 
 	/// Intervention end date (Unix timestamp)
 	///
 	/// - Mapped Dolibarr property: **datee**
-	public var dateEnd: Double?
+	public var dateEnd: Int?
 
 	/// Intervention duration in seconds
 	///
@@ -195,8 +195,8 @@ public final class DolibarrIntervention: CommonBusinessObject {
 		lines: [DolibarrInterventionLine] = [],
 		contractId: String? = nil,
 		projectId: String? = nil,
-		dateStart: Double? = nil,
-		dateEnd: Double? = nil,
+		dateStart: Int? = nil,
+		dateEnd: Int? = nil,
 		durationSeconds: String? = nil,
 		description: String? = nil,
 		clientRef: String? = nil,
